--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["flow", "react"],
-  "plugins": ["transform-class-properties"]
-}
