+++ src/components/Favourites.tsx
@@ -1,32 +1,30 @@
-/* 
+/*
  * This file is part of Flower.
- * 
+ *
  * Copyright ©2018 Nicolò Mazzucato
  * Copyright ©2018 Antonio Groza
  * Copyright ©2018 Brunello Simone
  * Copyright ©2018 Alessio Marotta
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- * 
+ *
  * Flower is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * Flower is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with Flower.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+import React, { Component } from "react";
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
+export class Favourites extends Component {
+  render() {
+    return <div></div>;
+  }
+}
