+++ src/index.css
@@ -26,4 +26,3 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
-
