+++ src/index.tsx
@@ -1,32 +1,29 @@
-/* 
+/*
  * This file is part of Flower.
- * 
+ *
  * Copyright ©2018 Nicolò Mazzucato
  * Copyright ©2018 Antonio Groza
  * Copyright ©2018 Brunello Simone
  * Copyright ©2018 Alessio Marotta
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- * 
+ *
  * Flower is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * Flower is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with Flower.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React, { Component } from 'react';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import App from "./App";
 
-export class Favourites extends Component {
-    render() {
-        return (
-            <div></div>
-        );
-    }
-}
+ReactDOM.render(<App />, document.getElementById("root"));
